% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_presentation.R
\name{addSmallLegend}
\alias{addSmallLegend}
\title{Change plot legend to make it smaller}
\usage{
addSmallLegend(myPlot, pointSize = 3, textSize = 6, spaceLegend = 0.1)
}
\arguments{
\item{myPlot}{ggplot2 object to edit}

\item{pointSize}{Point size}

\item{textSize}{Spacing between legend elements}
}
\value{
ggplot2 object
}
\description{
Change plot legend to make it smaller
}
\examples{
addSmallLegend(p)
}
