% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{check_file_extension}
\alias{check_file_extension}
\title{Check file extension and throw a warning/error
if it's not correct}
\usage{
check_file_extension(file, expected, abort = FALSE)
}
\arguments{
\item{file}{Path to file to check}

\item{expected}{Expected file extension}

\item{abort}{whether to abort if the extension
doesn't match}
}
\description{
Check file extension and throw a warning/error
if it's not correct
}
