% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{find_file}
\alias{find_file}
\title{Find a file with a given pattern in folder
\itemize{
\item Throws an error (or warning if abort is FALSE)
if it's not there
\item Throws a warning if multiple matches are found
}}
\usage{
find_file(folder, pattern, abort = FALSE, warning = TRUE)
}
\arguments{
\item{folder}{Folder to check}

\item{pattern}{Regex pattern}

\item{abort}{Whether to abort if no matches are found}

\item{warning}{Whether to send warning if no matches are found}
}
\value{
file path
}
\description{
Find a file with a given pattern in folder
\itemize{
\item Throws an error (or warning if abort is FALSE)
if it's not there
\item Throws a warning if multiple matches are found
}
}
