% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{geom_mean}
\alias{geom_mean}
\title{Geometric mean function for vectors with zeros}
\usage{
geom_mean(x)
}
\arguments{
\item{x}{Vector of positive values to compute the geometric mean on}
}
\value{
Numeric value
}
\description{
Geometric mean function for vectors with zeros
}
