% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{get_results_for_all_contrasts}
\alias{get_results_for_all_contrasts}
\title{Get result table from DESeq2 object between all levels vs reference
If the same OTU is selected for multiple comparisons, the padj and log2FoldChange
retained are the ones for the most significant comparison (lowest padj)}
\usage{
get_results_for_all_contrasts(diagdds, variable, comparisons)
}
\arguments{
\item{diagdds}{DESeq2 object}

\item{variable}{condition variable to test}

\item{comparisons}{n x 2 dataframe, each row is a comparison to perform}
}
\value{
result dataframe
}
\description{
Get result table from DESeq2 object between all levels vs reference
If the same OTU is selected for multiple comparisons, the padj and log2FoldChange
retained are the ones for the most significant comparison (lowest padj)
}
\examples{
get_results_for_all_results(diagdds, "Group", "A", c("B", "C"))
}
