% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-util.R
\name{get_taxa_annotations}
\alias{get_taxa_annotations}
\title{Format OTU metadata from a phyloseq object for the ComplexHeatmap package.}
\usage{
get_taxa_annotations(ps, levels = NULL, extra = NULL, colors = NULL)
}
\arguments{
\item{ps}{Phyloseq object}

\item{levels}{Taxonomic levels from tax table to display in the heatmap}

\item{extra}{(list) Extra columns (not in the tax table) to include.
The number of values must match the number of OTUs in the phyloseq object.}

\item{colors}{(list) Custom colors to use for any of the annotation variables.
Fields are variable names and values are named vectors
(names: factor levels, values: colors).}
}
\value{
List object with each variable as factor and an extra "col" field
with custom colors for each of those variable, as provided in the colors
parameter, plus custom colors for all of the taxonomic levels provided
}
\description{
Format OTU metadata from a phyloseq object for the ComplexHeatmap package.
}
\examples{
get_taxa_annotations(ps, c("Phylum", "Class"), extra=list(Nox=rep(F, ntaxa(ps))))
get_taxa_annotations(ps, "Phylum", extra=list(Nox=rep(F, ntaxa(ps))), 
                     cols=list(Nox=c(T="green", F="red")))
}
