% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{get_taxonomic_palette}
\alias{get_taxonomic_palette}
\title{Palette generator for taxonomic data
Uses database in github database repository}
\usage{
get_taxonomic_palette(taxa_names)
}
\arguments{
\item{taxa_names}{Vector of taxa names}
}
\value{
named vector, values are colors in HEX and names are taxa names
}
\description{
Palette generator for taxonomic data
Uses database in github database repository
}
\examples{
get_taxonomic_palette(c("Alphaproteobacteria", "Zetaproteobacteria", "Other"))
}
