% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_cmaiki}
\alias{load_cmaiki}
\title{Load outputs from C-MAIKI pipeline (MetaFlow|mics)
into phyloseq. You can either provide the output
folder from the pipeline run or the path to all
the individual files.}
\usage{
load_cmaiki(
  folder = NULL,
  abund_file = NULL,
  tax_file = NULL,
  meta_file = NULL,
  tree_file = NULL,
  otu_id = 100
)
}
\arguments{
\item{folder}{Output folder from MetaFlow|mics}

\item{abund_file}{Path to abundance file
(either .count_table or .shared)}

\item{tax_file}{Path to .cons.taxonomy file}

\item{meta_file}{Path to sample metadata (csv formatted).}

\item{tree_file}{Path to phylogenetic tree (otpional)
Assumes there is a header and the first column are the sample names}

\item{otu_id}{OTU clustering threshold in case
multiple values where used}
}
\value{
phyloseq object
}
\description{
Load outputs from C-MAIKI pipeline (MetaFlow|mics)
into phyloseq. You can either provide the output
folder from the pipeline run or the path to all
the individual files.
}
\examples{
load_cmaiki("pipeline-2022-07-9/", )
load_cmaiki(abund_file="OTUs.100.rep.count_table", tax_file="OTUs.100.cons.taxonomy", meta_file="proj/metadata.csv")
}
