% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_constaxonomy}
\alias{load_constaxonomy}
\title{Load consensus taxonomy from mothur}
\usage{
load_constaxonomy(file)
}
\arguments{
\item{file}{Path to cons.taxonomy file}
}
\value{
Taxonomy dataframe (OTU x taxonomic ranks)
}
\description{
Load consensus taxonomy from mothur
}
\examples{
load_constaxonomy("annotations.100.cons.taxonomy")
}
