% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_counttable}
\alias{load_counttable}
\title{Load OTU abundance file from mothur (count table format)}
\usage{
load_counttable(file)
}
\arguments{
\item{file}{Path to count_table file}
}
\value{
Abundance dataframe (OTU x sample)
}
\description{
Load OTU abundance file from mothur (count table format)
}
\examples{
load_counttable("OTUs.100.rep.count_table")
}
