% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_metadata}
\alias{load_metadata}
\title{Load metadata file}
\usage{
load_metadata(file)
}
\arguments{
\item{file}{Path to metadata file}
}
\value{
Data Frame
}
\description{
Load metadata file
}
\examples{
load_metadata("metadata.csv")
}
