% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_shared}
\alias{load_shared}
\title{Load OTU abundance file from mothur (shared format)}
\usage{
load_shared(file)
}
\arguments{
\item{file}{Path to shared file}
}
\value{
Abundance dataframe (sample x OTU)
}
\description{
Load OTU abundance file from mothur (shared format)
}
\examples{
load_shared("abundance.100.shared")
}
