% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{merge_samples3}
\alias{merge_samples3}
\title{Wrapper around speedyseq::merge_samples2 for combination of factors}
\usage{
merge_samples3(ps, groups)
}
\arguments{
\item{ps}{Phyloseq object}

\item{groups}{variables to define sample groups}
}
\value{
Phyloseq object with sample merged
}
\description{
Wrapper around speedyseq::merge_samples2 for combination of factors
}
\examples{
merge_samples3(ps, c("Site", "Season"))
}
