% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{plot_ordination_with_arrows}
\alias{plot_ordination_with_arrows}
\title{Plot of constrained ordination plot with environment variable effect shown as arrows}
\usage{
plot_ordination_with_arrows(
  data,
  arrows,
  x = "CCA1",
  y = "CCA2",
  fill = "black",
  size = 3,
  arrow_rescale = 0.5,
  ellipse = TRUE,
  ...
)
}
\arguments{
\item{data}{Ordination dataframe with metadata (long format for components)}

\item{arrows}{Environment factor (arrows) to overlay}

\item{x}{Variable name for x-axis}

\item{y}{Variable name for y-axis}

\item{fill}{Variable name for points fill color}

\item{size}{Point size}

\item{arrow_rescale}{Rescale all arrows with by a factor}

\item{ellipse}{Whether to draw ellipses around fill groups}

\item{...}{Additional params to pass to geom_point()}
}
\value{
ggplot2 object
}
\description{
Plot of constrained ordination plot with environment variable effect shown as arrows
}
\examples{
plot_ordination_with_arrows(df, arrows, "NMDS1", "NMDS2", fill="condition")
}
