% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{run_adonis}
\alias{run_adonis}
\title{Wrapper around vegan::adonis}
\usage{
run_adonis(ps, meta_vars, ...)
}
\arguments{
\item{ps}{Phyloseq object}

\item{meta_vars}{Variable(s) to test}

\item{...}{Extra parameters to pass to the adonis function}
}
\value{
dataframe with r2 and pvalue for PERMANOVA test
}
\description{
Wrapper around vegan::adonis
}
\examples{
run_adonis(ps, "Group+Season", permutations=999)
}
