% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Wrapper around DESeq functions}
\usage{
run_deseq2(ps, design, gm = FALSE, pseudocount = 0)
}
\arguments{
\item{ps}{Phyloseq object}

\item{design}{Experimental design. The last variable needs to be the experimental condition to test for}

\item{gm}{Whether or not to use geometric mean for size factor estimation. Might be necessary if there are a lot of zero (or add pseudocount}

\item{pseudocount}{Pseudocount to add to abundance table if there are too many zeros}
}
\value{
DESeq2 object
}
\description{
Wrapper around DESeq functions
}
\examples{
run_deseq2(ps, ~ Season + Group, gm=TRUE)
run_deseq2(ps, ~ Season + Group, pseudocount=1)
}
