% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{run_ordination}
\alias{run_ordination}
\title{Wrapper around phyloseq's ordinate function
Adds the metadata to the component scores}
\usage{
run_ordination(ps, method = "NMDS", ...)
}
\arguments{
\item{ps}{Phyloseq object. Needs to contain a phylogenetic tree if the distance matrix is unifrac-based}

\item{method}{Ordinaton method. See \code{ordinate()} function to see all the available methods}

\item{...}{Extra arguments to provide to the underlying ordination function}
}
\value{
List with ordination model, dataframe with first 2 components and metadata information, and biplot data if supervised
}
\description{
Wrapper around phyloseq's ordinate function
Adds the metadata to the component scores
}
\examples{
run_ordination(ps, method="NMDS", distance="bray", trymax=300)
run_ordination(ps, method="PCoA")
run_ordination(ps, "CCA", formula=~ DO + pH + SPC)
}
