% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{taxa_barplot}
\alias{taxa_barplot}
\title{Taxonomic stacked barplot of phyloseq object.}
\usage{
taxa_barplot(
  ps,
  x = NULL,
  y = "Abundance",
  taxrank = "Class",
  rows = NULL,
  cols = NULL,
  min_relabund = 0.01,
  return_df = FALSE,
  nrows_legend = 20
)
}
\arguments{
\item{ps}{Phyloseq object to plot}

\item{x}{Categorical variable in metadata, displayed in x-axis}

\item{y}{Variable for the height of bars. Either "Abundance" or
"relabund" (the latter normalize the bars to the same height)}

\item{taxrank}{Taxonomic rank to gather taxa (stacked for each level of x)}

\item{rows}{Factor for row facets}

\item{cols}{Factor for col facets}

\item{min_relabund}{Below this threshold, replace taxa name to "< x\%" to
make the plot for readable and reduce the number of colors}

\item{return_df}{Whether to return the formatted dataframe or plot}
}
\value{
ggplot object (geom_bar) or a dataframe if return_df is TRUE
}
\description{
Taxonomic stacked barplot of phyloseq object.
}
\examples{
taxa_barplot(ps, x="sample_type", taxrank="Phylum", min_relabund=0.01)
taxa_barplot(ps, x="sample_type", y="relabund", taxrank="Phylum", rows="Season")
}
