% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_presentation.R
\name{theme_presentation}
\alias{theme_presentation}
\title{Theme for black background presentations}
\usage{
theme_presentation(tsize = 24)
}
\arguments{
\item{tsize}{Base text size}
}
\value{
ggplot2 theme object
}
\description{
Theme for black background presentations
}
\examples{
theme_presentation(20)
}
