% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_presentation.R
\name{theme_report}
\alias{theme_report}
\title{Theme for reports. Starts with theme_grey and then modify some parts}
\usage{
theme_report(tsize = 18, xaxis = 90, xhjust = 1)
}
\arguments{
\item{tsize}{Base text size}

\item{xaxis}{Rotation for x-axis text}

\item{xhjust}{Horizontal text justification (from 0: left to 1: right)}
}
\value{
ggplot2 theme object
}
\description{
Theme for reports. Starts with theme_grey and then modify some parts
}
\examples{
theme_report(15, 45, 1)
}
